requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
schedule==1.2.1
slack-sdk==3.27.0
html5lib==1.1
lxml==5.2.1
python-dateutil==2.8.2
retrying==1.3.3
pdfminer.six==20221105